feat: extract fusion into lib and wrap it with NAT
