+++ services/agent/src/vss_agents/tools/search.py
@@ -42,31 +42,31 @@ from pydantic import ConfigDict
 from pydantic import Field
 from pydantic import model_validator
 
+from lib.fusion import FusedSegment
+from lib.fusion import FusionInput
+from lib.fusion import FusionMethod
+from lib.fusion import FusionOutput
+from lib.fusion.ranking_models import DEFAULT_CHUNK_SECONDS as _CHUNK_SECONDS
+from lib.fusion.ranking_models import ChunkKey
+from lib.fusion.ranking_models import EmbeddingSpaceName
+from lib.fusion.ranking_models import FusableSearchOutput
+from lib.fusion.ranking_models import RankedList
+from lib.utils.time_convert import datetime_to_iso8601
+from lib.utils.time_convert import iso8601_to_datetime
 from vss_agents.agents.data_models import AgentMessageChunk
 from vss_agents.agents.data_models import AgentMessageChunkType
-from vss_agents.data_models.ranking import DEFAULT_CHUNK_SECONDS as _CHUNK_SECONDS
-from vss_agents.data_models.ranking import ChunkKey
-from vss_agents.data_models.ranking import EmbeddingSpaceName
-from vss_agents.data_models.ranking import FusableSearchOutput
-from vss_agents.data_models.ranking import RankedList
 from vss_agents.data_models.search import CriticResult
 from vss_agents.data_models.search import DecomposedQuery
 from vss_agents.data_models.search import SearchInput
 from vss_agents.data_models.search import SearchResult
 from vss_agents.tools.attribute_search import DEFAULT_BEHAVIOR_INDEX
 from vss_agents.tools.embed_search import EmbedSearchOutput
-from vss_agents.tools.fusion import FusedSegment
-from vss_agents.tools.fusion import FusionInput
-from vss_agents.tools.fusion import FusionMethod
-from vss_agents.tools.fusion import FusionOutput
 from vss_agents.tools.spaces_registry import ANCHOR_EMBEDDING_SPACE
 from vss_agents.tools.spaces_registry import EMBEDDING_SPACE_ADAPTERS
 from vss_agents.tools.vst.utils import get_streams_info
 from vss_agents.utils.es_client import VSSESClient
 from vss_agents.utils.reasoning_utils import get_llm_reasoning_bind_kwargs
 from vss_agents.utils.reasoning_utils import get_thinking_tag
-from vss_agents.utils.time_convert import datetime_to_iso8601
-from vss_agents.utils.time_convert import iso8601_to_datetime
 from vss_agents.utils.time_measure import TimeMeasure
 
 logger = logging.getLogger(__name__)
